Moving virtual_overrider_self_life_support to separate include file, with iwyu cleanup.
