+++ tests/extra_python_package/test_files.py
@@ -48,6 +48,7 @@ detail_headers = {
     "include/pybind11/detail/smart_holder_type_casters.h",
     "include/pybind11/detail/type_caster_base.h",
     "include/pybind11/detail/typeid.h",
+    "include/pybind11/detail/virtual_overrider_self_life_support.h",
 }
 
 cmake_files = {
